Split up code into one file per class
