+++ esphome/components/number/number.h
@@ -3,6 +3,8 @@
 #include "esphome/core/component.h"
 #include "esphome/core/entity_base.h"
 #include "esphome/core/helpers.h"
+#include "number_call.h"
+#include "number_traits.h"
 
 namespace esphome {
 namespace number {
@@ -19,54 +21,6 @@ namespace number {
   }
 
 class Number;
-
-class NumberCall {
- public:
-  explicit NumberCall(Number *parent) : parent_(parent) {}
-  void perform();
-
-  NumberCall &set_value(float value) {
-    value_ = value;
-    return *this;
-  }
-  const optional<float> &get_value() const { return value_; }
-
- protected:
-  Number *const parent_;
-  optional<float> value_;
-};
-
-enum NumberMode : uint8_t {
-  NUMBER_MODE_AUTO = 0,
-  NUMBER_MODE_BOX = 1,
-  NUMBER_MODE_SLIDER = 2,
-};
-
-class NumberTraits {
- public:
-  void set_min_value(float min_value) { min_value_ = min_value; }
-  float get_min_value() const { return min_value_; }
-  void set_max_value(float max_value) { max_value_ = max_value; }
-  float get_max_value() const { return max_value_; }
-  void set_step(float step) { step_ = step; }
-  float get_step() const { return step_; }
-
-  /// Get the unit of measurement, using the manual override if set.
-  std::string get_unit_of_measurement();
-  /// Manually set the unit of measurement.
-  void set_unit_of_measurement(const std::string &unit_of_measurement);
-
-  // Get/set the frontend mode.
-  NumberMode get_mode() const { return this->mode_; }
-  void set_mode(NumberMode mode) { this->mode_ = mode; }
-
- protected:
-  float min_value_ = NAN;
-  float max_value_ = NAN;
-  float step_ = NAN;
-  optional<std::string> unit_of_measurement_;  ///< Unit of measurement override
-  NumberMode mode_{NUMBER_MODE_AUTO};
-};
 
 /** Base-class for all numbers.
  *
